+++ docs/examples/robot_pots_1.py
@@ -4,9 +4,9 @@ from signal import pause
 
 robot = Robot(left=(4, 14), right=(17, 18))
 
-left = MCP3008(0)
-right = MCP3008(1)
+left_pot = MCP3008(0)
+right_pot = MCP3008(1)
 
-robot.source = zip_values(left, right)
+robot.source = zip_values(left_pot, right_pot)
 
 pause()