+++ docs/examples/sin_values.py
@@ -1,0 +1,13 @@
+from gpiozero import Motor, Servo, TonalBuzzer
+from gpiozero.tools import sin_values
+from signal import pause
+
+motor = Motor(2, 3)
+servo = Servo(4)
+buzzer = TonalBuzzer(5)
+
+motor.source = sin_values()
+servo.source = motor
+buzzer.source = motor
+
+pause()